data_product/requirements.txt
arrow>=1.2.2
delta-spark>=2.4.0
deprecated>=1.2.13
importlib-metadata==4.12.0
py4j>=0.10.9.5
pyspark==3.4.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pytest>=7.4.0
click>=8.1.6